Part 4 and Part 5: add consistent-hash & add distributed node & add demo (main.go)

+++ TinyCache/lru/lru_test.go
@@ -12,7 +12,7 @@ func (d String) Len() int {
 }
 
 func TestGet(t *testing.T) {
-	lru := NewCache(int64(0), nil)
+	lru := NewCache(0, nil)
 	lru.Add("key1", String("1234"))
 	if v, ok := lru.Get("key1"); !ok || string(v.(String)) != "1234" {
 		t.Fatal("cache hit key1=1234 failed")
@@ -26,7 +26,7 @@ func TestRemoveOldest(t *testing.T) {
 	k1, k2, k3 := "key1", "key2", "k3"
 	v1, v2, v3 := "value1", "value2", "v3"
 	cap := len(k1 + k2 + v1 + v2)
-	lru := NewCache(int64(cap), nil)
+	lru := NewCache(uint32(int64(cap)), nil)
 	lru.Add(k1, String(v1))
 	lru.Add(k2, String(v2))
 	lru.Add(k3, String(v3))
@@ -41,7 +41,7 @@ func TestOnEvicted(t *testing.T) {
 	callback := func(key string, value Value) {
 		keys = append(keys, key)
 	}
-	lru := NewCache(int64(12), callback)
+	lru := NewCache(12, callback)
 	lru.Add("key1", String("123456"))
 	lru.Add("k2", String("k2"))
 	lru.Add("k3", String("k3"))
